+++ packages/core/tree/src/index.js
@@ -4,6 +4,13 @@ export { default } from './components/Tree';
 
 export type { RenderItemParams } from './components/TreeItem/TreeItem-types';
 
-export type { ItemId, Path, TreeData, TreeItem, TreePosition } from './types';
+export type {
+  ItemId,
+  Path,
+  TreeData,
+  TreeItem,
+  TreeSourcePosition,
+  TreeDestinationPosition,
+} from './types';
 
 export { mutateTree, moveItemOnTree } from './utils/tree';