+++ packages/editor/editor-core/src/__tests__/integration/message-renderer/mention-2.ts
@@ -21,7 +21,7 @@ BrowserTestCase(
   'Mention: user should see picker if they type "@"',
   { skip: ['ie'] },
   async client => {
-    const browser = await new Page(client);
+    const browser = new Page(client);
     await browser.goto(messageEditor);
     await browser.waitForSelector(editable);
     await browser.type(editable, '@');
@@ -35,7 +35,7 @@ BrowserTestCase(
   'Mention: text@ should not invoke picker',
   { skip: ['ie'] },
   async client => {
-    const browser = await new Page(client);
+    const browser = new Page(client);
     await browser.goto(messageEditor);
     await browser.waitForSelector(editable);
     await browser.type(editable, 'test@');
@@ -47,7 +47,7 @@ BrowserTestCase(
   'Mention: user should be able remove mention on backspace',
   { skip: ['safari', 'ie'] },
   async client => {
-    const browser = await new Page(client);
+    const browser = new Page(client);
     await browser.goto(messageEditor);
     await browser.waitForSelector(editable);
     await insertMention(browser, 'Carolyn');
@@ -64,7 +64,7 @@ BrowserTestCase(
   'Mention: @ <space> should not invoke picker',
   { skip: ['ie'] },
   async client => {
-    const browser = await new Page(client);
+    const browser = new Page(client);
     await browser.goto(messageEditor);
     await browser.waitForSelector(editable);
     await browser.type(editable, '@ Carolyn');
@@ -76,7 +76,7 @@ BrowserTestCase(
   'Mention: insert on space if unique exact nickname match, with multiple results',
   { skip: ['ie', 'edge'] },
   async client => {
-    const browser = await new Page(client);
+    const browser = new Page(client);
     await browser.goto(messageEditor);
     await browser.waitForSelector(editable);
     await browser.type(editable, '@');
@@ -94,7 +94,7 @@ BrowserTestCase(
   'Mention: user should see space after node',
   { skip: ['safari', 'ie'] },
   async client => {
-    const browser = await new Page(client);
+    const browser = new Page(client);
     await browser.waitForSelector(editable);
     await browser.goto(messageEditor);
     await insertMention(browser, 'Summer');
@@ -108,7 +108,7 @@ BrowserTestCase(
   'Mention: escape closes picker',
   { skip: ['safari', 'ie'] },
   async client => {
-    const browser = await new Page(client);
+    const browser = new Page(client);
     await browser.goto(messageEditor);
     await browser.waitForSelector(editable);
     await browser.type(editable, '@');