+++ packages/core/navigation-next/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @atlaskit/navigation-next
 
+## 3.16.0
+- [minor] introduce flyout on hover, with feature-flag property [5c00034](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/5c00034)
+
+## 3.15.6
+- [patch] Avoid re-mount of custom components on re-render [1a47c0b](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/1a47c0b)
+
+## 3.15.5
+- [patch] Updated dependencies [65c6514](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/65c6514)
+  - @atlaskit/docs@5.0.8
+  - @atlaskit/avatar@14.0.11
+  - @atlaskit/button@9.0.13
+  - @atlaskit/drawer@2.1.1
+  - @atlaskit/dropdown-menu@6.1.17
+  - @atlaskit/field-base@11.0.8
+  - @atlaskit/global-navigation@5.0.1
+  - @atlaskit/modal-dialog@7.0.2
+  - @atlaskit/onboarding@6.0.2
+  - @atlaskit/section-message@1.0.8
+  - @atlaskit/select@6.0.2
+  - @atlaskit/toggle@5.0.9
+  - @atlaskit/tooltip@12.1.1
+  - @atlaskit/icon@14.0.0
+
 ## 3.15.4
 - [patch] Updated dependencies [ac88888](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ac88888)
   - @atlaskit/global-navigation@5.0.0