+++ packages/core/portal/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @atlaskit/portal
 
+## 0.0.10
+- [patch] Updated dependencies [65c6514](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/65c6514)
+  - @atlaskit/docs@5.0.8
+  - @atlaskit/button@9.0.13
+  - @atlaskit/flag@9.0.11
+  - @atlaskit/inline-dialog@9.0.6
+  - @atlaskit/modal-dialog@7.0.2
+  - @atlaskit/onboarding@6.0.2
+  - @atlaskit/tooltip@12.1.1
+  - @atlaskit/icon@14.0.0
+
 ## 0.0.9
 - [patch] Updated dependencies [d5a043a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/d5a043a)
   - @atlaskit/icon@13.8.1