+++ packages/elements/task-decision/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @atlaskit/task-decision
 
+## 9.0.2
+- [patch] Async load avatargroup because it also pulls in dropdown [fb13496](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/fb13496)
+
+## 9.0.1
+- [patch] Updated dependencies [65c6514](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/65c6514)
+  - @atlaskit/docs@5.0.8
+  - @atlaskit/avatar@14.0.11
+  - @atlaskit/avatar-group@2.1.3
+  - @atlaskit/button@9.0.13
+  - @atlaskit/editor-common@17.0.7
+  - @atlaskit/renderer@24.2.1
+  - @atlaskit/icon@14.0.0
+
 ## 9.0.0
 - [major] Update RXJS dependency to ^5.5.0 [927ae63](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/927ae63)