+++ packages/growth/feature-flag-client/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/feature-flag-client
 
+## 1.0.5
+- [patch] MEP-211: Supporting "value" attribute [8c0ddfc](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/8c0ddfc)
+
 ## 1.0.4
 - [patch] MEP-103: Allowing simple flags to have strings or booleans as values [4e6f8ab](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/4e6f8ab)