+++ packages/people-and-teams/profilecard/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @atlaskit/profilecard
 
+## 4.0.10
+- [patch] Updated dependencies [65c6514](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/65c6514)
+  - @atlaskit/docs@5.0.8
+  - @atlaskit/avatar@14.0.11
+  - @atlaskit/button@9.0.13
+  - @atlaskit/icon@14.0.0
+
 ## 4.0.9
 - [patch] Remove promise polyfill to fix Jira SSR [e793f6e](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/e793f6e)