+++ packages/search/global-search/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @atlaskit/global-search
 
+## 5.9.5
+- [patch] fix higlight problem due to duplicate react keys [beeba8a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/beeba8a)
+
+## 5.9.4
+- [patch] icon-file-type and icon-object publish glyphs, svgs, and es5 instead of just dist [0823d35](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/0823d35)
+
+## 5.9.3
+- [patch] Exposure event gets trigger on initial load instead when search results return [b89cda9](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/b89cda9)
+
+## 5.9.2
+- [patch] Trim search input [488257d](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/488257d)
+
+## 5.9.1
+- [patch] Updated dependencies [709b239](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/709b239)
+- [patch] Updated dependencies [65c6514](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/65c6514)
+  - @atlaskit/icon@14.0.0
+  - @atlaskit/icon-file-type@1.0.0
+  - @atlaskit/icon-object@1.0.0
+  - @atlaskit/docs@5.0.8
+  - @atlaskit/avatar@14.0.11
+  - @atlaskit/button@9.0.13
+  - @atlaskit/dropdown-menu@6.1.17
+  - @atlaskit/flag@9.0.11
+  - @atlaskit/modal-dialog@7.0.2
+  - @atlaskit/navigation@33.1.11
+  - @atlaskit/quick-search@4.2.9
+
 ## 5.9.0
 - [minor] Feature to use quicknav for people search [9f1d252](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/9f1d252)